64bit safe bmp texture loading on Mac.

+++ src/texture.h
@@ -42,35 +42,35 @@ public:
 
     struct BMPFILEHEADER
     {
-        short   bfType;
-        long    bfSize;
-        short   bfReserved1;
-        short   bfReserved2;
-        long    bfOffBits;
+        int16_t   bfType;
+        int32_t   bfSize;
+        int16_t   bfReserved1;
+        int16_t   bfReserved2;
+        int32_t    bfOffBits;
     };
 
     struct BMPINFOHEADER
     {
-        long    biSize;
-        long    biWidth;
-        long    biHeight;
-        short   biPlanes;
-        short   biBitCount;
-        long    biCompression;
-        long    biSizeImage;
-        long    biXPelsPerMeter;
-        long    biYPelsPerMeter;
-        long    biClrUsed;
-        long    biClrImportant;
+        int32_t    biSize;
+        int32_t    biWidth;
+        int32_t    biHeight;
+        int16_t   biPlanes;
+        int16_t   biBitCount;
+        int32_t    biCompression;
+        int32_t    biSizeImage;
+        int32_t    biXPelsPerMeter;
+        int32_t    biYPelsPerMeter;
+        int32_t    biClrUsed;
+        int32_t    biClrImportant;
     };
 
     struct IMAGEDATA
     {
         std::vector<unsigned char> pData;
-        long    Width;
-        long    Height;
-        long    Padding;
-        short   Channels;
+        int32_t    Width;
+        int32_t    Height;
+        int32_t    Padding;
+        int16_t   Channels;
         unsigned int bpp;
     };
